Remove: api/hooks 내부의 불필요한 hook을 제거
