+++ src/components/player/SoccerPlayer/SoccerPlayerName.tsx
@@ -1,6 +1,7 @@
-import { useSoccerPlayerName } from 'src/pages/api/hooks/query/useSoccerPlayerName';
+import { useCaseMatchSearch } from 'src/useCases/useCaseMatchSearch';
 
 export const SoccerPlayerName = ({ spId }: { spId: number }) => {
+  const { useSoccerPlayerName } = useCaseMatchSearch();
   const name = useSoccerPlayerName(spId);
 
   const lastName = name?.result.split(' ')[1] || name?.result.split(' ')[0];
