+++ src/constants/matchMeta.ts
@@ -1,0 +1,38 @@
+export const matchMetaData = [
+  {
+    matchtype: 30,
+    desc: '리그 친선',
+  },
+  {
+    matchtype: 40,
+    desc: '클래식 1on1',
+  },
+  {
+    matchtype: 50,
+    desc: '공식경기',
+  },
+  {
+    matchtype: 52,
+    desc: '감독모드',
+  },
+  {
+    matchtype: 60,
+    desc: '공식 친선',
+  },
+  {
+    matchtype: 204,
+    desc: '볼타 친선',
+  },
+  {
+    matchtype: 214,
+    desc: '볼타 공식',
+  },
+  {
+    matchtype: 224,
+    desc: '볼타 AI대전',
+  },
+  {
+    matchtype: 234,
+    desc: '볼타 커스텀',
+  },
+];
