--- src/pages/api/user/getUserAccessId.ts
@@ -1,5 +1,0 @@
-export interface UserProfile {
-  accessId: string; //	유저 고유 식별자
-  nickname: string; //	유저 닉네임
-  level: number; //	유저 레벨
-}
