+++ src/useCases/changeServerDataIntoRenderData.ts
@@ -33,7 +33,6 @@ export const changeServerDataIntoRenderData = (
   matchDetailData: IMatchDetailData,
   userNickName: string,
 ) => {
-  // 여기서 matchDetailData가 예상과 다르게 들어와도 default 상태를 정의해주면 문제없이 동작할 것으로 보인다.
   const newState: IViewData = {
     matchType: '',
     matchResult: '승',
@@ -41,7 +40,7 @@ export const changeServerDataIntoRenderData = (
     leftPlayer: {
       nickName: '',
       goalCount: 0,
-      // 🟡: 유저의 닉네임이 들어갈만한 최소 공간(점유율)
+      // 유저의 닉네임이 들어갈만한 최소 공간(점유율)
       possession: 70,
       bestPlayer: {
         spPosition: '',
@@ -50,10 +49,9 @@ export const changeServerDataIntoRenderData = (
       },
     },
     rightPlayer: {
-      // 🟡: string type의 defualt 닉네임
       nickName: '무명유저',
       goalCount: 0,
-      // 🟡: 유저의 닉네임이 들어갈만한 최소 공간(점유율)
+      // 유저의 닉네임이 들어갈만한 최소 공간(점유율)
       possession: 30,
       bestPlayer: {
         spPosition: '',
@@ -64,20 +62,12 @@ export const changeServerDataIntoRenderData = (
     matchDetails: [
       {},
       {
-        // 🟡: MatchInfo[] 형식의 defautl data를 넣어줌
         player: [{}],
       },
     ] as MatchInfo[],
   };
   // 시간변환
   newState.matchDate = changeDateUtil(matchDetailData.matchDate);
-  // leftPlyaer, rightPlaery
-  // left: 검색한 본인
-
-  // 기권패인 경우에는 관련 값들이 빈상태로 오는 경우도 존재한다.
-  // 몰수패는 어떻게 보여줄꺼야....
-
-  // 상대방이 특정 이유(닉네임변경, 계정삭제)로 인해 데이터가 넘어오지 않을 경우 처리
   if (matchDetailData.matchInfo.length === 1) {
     const searcherData = matchDetailData.matchInfo[0];
     const bestPlayer = pickBestPlayer(searcherData);
