+++ src/useCases/useCaseMatchSearch.ts
@@ -1,11 +1,10 @@
 // 검색시 필요한 query를 모두 넣어주자.
 
 import { QueryClient, useInfiniteQuery, useQueryClient } from 'react-query';
-import { metaQueryKey } from 'src/pages/api/hooks/useGetMetaQuery';
-// import { metaQueryKey } from 'src/hooks/useGetMetaQuery';
 
 import { matchAPI } from 'src/pages/api/player';
 import type { AccessId, IMetaSpId } from 'src/pages/api/player/type';
+import { metaQueryKey } from 'src/useCases/useCaseGetMetaData';
 import { getSoccerPlayerNameBySpId } from './getSoccerPlayerNameBySpId';
 
 export const useCaseMatchSearch = () => {
